Moved from assert statement for corpus data validation check to print statements to allow identification of all the corpus data issues. This will allow identification of systemic issues with the graph database.
Also created a flag to validate the corpus data against the corpus schema (what was there previously) and to also check to see if the corpus data has elements that are not identified in the corpus schema yet.

Fixed corpus_data.json to run with the code - but note that there are still issues with the graph consistency right now.

Updated the corpus_schema.json to include BatteryController BEAM, BEAM_CAP, SERVO, and FuselageNACA classes.  Left Instrument_Battery out for now.